Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -4,7 +4,7 @@ int main()
     int r;
     scanf("%r",&r);
     float pi=3.14;
-    int area=(pi*r*r);
-    printf("%d",area);
+    float area=(pi*r*r);
+    printf("%.2f",area);
     return 0;
 }
